Move unused stats folder to sandbox
